requirements.txt
aiogram==2.25.1
python-dotenv==1.0.0
psutil==5.9.5
nest-asyncio==1.5.8
